code/04-mcp-server/requirements.txt
anthropic>=0.40.0
mcp>=1.0.0
python-dotenv>=1.0.0
